Testing/requirements.txt
pyserial>=3.5
colorama>=0.4.6